+++ changelog.md
@@ -2,6 +2,7 @@
 - Added different sounds for blocks of different materials (BluSunrize)
 - Added "forge:ores" tag to all IE ores (Malte)
 - Re-Added packing and unpacking recipes for the metal press (BluSunrize)
+- Added cloche recipe to grow bamboo (BluSunrize)
 - Fixed bounding boxes for vertical covered conveyors (Malte)
 - Fixed metal press not forming when mirrored (Malte)
 - Fixed extracting conveyors not saving their extract direction (BluSunrize)